chore: cleanup everything
