+++ lib/logic/component_index_service.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 
-base class ComponentsIndexService {
+base class ComponentIndexService {
   final state = ValueNotifier<Set<int>>({});
 
   void add(int index) => state.value = {...state.value..add(index)};
@@ -12,6 +12,6 @@ base class ComponentsIndexService {
   bool contains(int index) => state.value.contains(index);
 }
 
-final class Selected extends ComponentsIndexService {}
+final class Selected extends ComponentIndexService {}
 
-final class Hovered extends ComponentsIndexService {}
+final class Hovered extends ComponentIndexService {}
