+++ lib/main.dart
@@ -37,6 +37,8 @@ part 'widgets/right_sidebar.dart';
 
 part 'widgets/top_bar.dart';
 
+part 'widgets/home_page.dart';
+
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await windowManager.ensureInitialized();
@@ -46,9 +48,7 @@ void main() async {
   if (Platform.isMacOS) {
     await WindowManipulator.initialize(enableWindowDelegate: true);
     final delegate = _MyDelegate();
-    WindowManipulator.addNSWindowDelegate(
-      delegate,
-    );
+    WindowManipulator.addNSWindowDelegate(delegate);
     final options = NSAppPresentationOptions.from({
       NSAppPresentationOption.fullScreen,
       NSAppPresentationOption.autoHideToolbar,
@@ -93,10 +93,10 @@ class _MyDelegate extends NSWindowDelegate {
 void setup() {
   final register = GetIt.I.registerSingleton;
 
-  register<Components>(Components());
+  register<ComponentService>(ComponentService());
   register<Selected>(Selected());
   register<Hovered>(Hovered());
-  register<CanvasState>(CanvasState());
+  register<CanvasService>(CanvasService());
 }
 
 class Main extends StatelessWidget with GetItMixin {
@@ -131,7 +131,8 @@ class Main extends StatelessWidget with GetItMixin {
                       type: PlatformProvidedMenuItemType.zoomWindow,
                     ),
                     PlatformProvidedMenuItem(
-                        type: PlatformProvidedMenuItemType.hide),
+                      type: PlatformProvidedMenuItemType.hide,
+                    ),
                     PlatformProvidedMenuItem(
                       type: PlatformProvidedMenuItemType.hideOtherApplications,
                     ),
@@ -139,7 +140,8 @@ class Main extends StatelessWidget with GetItMixin {
                       type: PlatformProvidedMenuItemType.toggleFullScreen,
                     ),
                     PlatformProvidedMenuItem(
-                        type: PlatformProvidedMenuItemType.quit),
+                      type: PlatformProvidedMenuItemType.quit,
+                    ),
                   ]),
                 ],
               ),
@@ -248,15 +250,18 @@ class Main extends StatelessWidget with GetItMixin {
           ),
           useMaterial3: true,
         ),
-        home: MultiBlocProvider(providers: [
-          BlocProvider(create: (_) => CanvasEventsCubit()),
-          BlocProvider(create: (_) => CanvasTransformCubit()),
-          BlocProvider(create: (_) => KeysCubit()),
-          BlocProvider(create: (_) => PointerCubit(Offset.zero)),
-          BlocProvider(create: (_) => ToolCubit(ToolType.move)),
-          //
-          BlocProvider(create: (_) => DebugPointCubit()),
-        ], child: HomePage()),
+        home: MultiBlocProvider(
+          providers: [
+            BlocProvider(create: (_) => CanvasEventsCubit()),
+            BlocProvider(create: (_) => CanvasTransformCubit()),
+            BlocProvider(create: (_) => KeysCubit()),
+            BlocProvider(create: (_) => PointerCubit(Offset.zero)),
+            BlocProvider(create: (_) => ToolCubit(ToolType.move)),
+            //
+            BlocProvider(create: (_) => DebugPointCubit()),
+          ],
+          child: HomePage(),
+        ),
       ),
     );
   }
@@ -274,772 +279,4 @@ void handleGoForward() {
     return;
   }
   componentsNotifier.reorder(index, index + 1);
-}
-
-class HomePage extends StatefulWidget with GetItStatefulWidgetMixin {
-  HomePage({
-    super.key,
-  });
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> with GetItStateMixin {
-  final oMatrix = ValueNotifier(Matrix4.identity());
-  var isShowColorPicker = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final canvasStateProvider =
-        watchX((CanvasState canvasState) => canvasState.state);
-    //
-    final canvasEvents = context.watch<CanvasEventsCubit>();
-
-    final leftClick = canvasEvents.state.contains(CanvasEvent.leftClick);
-
-    final middleClick = canvasEvents.state.contains(CanvasEvent.middleClick);
-
-    final isCreateTooling = canvasEvents.containsAny({
-      CanvasEvent.creatingRectangle,
-      CanvasEvent.creatingFrame,
-      CanvasEvent.creatingText,
-    });
-    final isComponentTooling = canvasEvents.containsAny({
-      CanvasEvent.draggingComponent,
-      CanvasEvent.resizingComponent,
-      CanvasEvent.rotatingComponent,
-    });
-    final isNotCanvasTooling = isComponentTooling || isCreateTooling;
-    final isCanvasTooling = canvasEvents.containsAny({
-      CanvasEvent.panningCanvas,
-      CanvasEvent.zoomingCanvas,
-    });
-    final isZooming = canvasEvents.containsAny(
-      {CanvasEvent.zoomingCanvas},
-    );
-
-    final transformationController =
-        context.watch<CanvasTransformCubit>().state;
-    final tool = context.watch<ToolCubit>().state;
-    final isToolHand = tool == ToolType.hand;
-    //
-    final components = watchX((Components components) => components.state);
-    final selected = watchX((Selected selected) => selected.state);
-    final hovered = watchX((Hovered hovered) => hovered.state);
-    //
-    final keys = context.watch<KeysCubit>().state;
-    keys;
-    final pressedMeta = keys.contains(LogicalKeyboardKey.metaLeft) ||
-        keys.contains(LogicalKeyboardKey.metaRight) ||
-        keys.contains(LogicalKeyboardKey.meta);
-
-    final mqSize = MediaQuery.of(context).size;
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return RawKeyboardListener(
-      focusNode: FocusNode(),
-      autofocus: true,
-      onKey: (event) {
-        event.isKeyPressed(event.logicalKey)
-            ? context.read<KeysCubit>().add(event.logicalKey)
-            : context.read<KeysCubit>().remove(event.logicalKey);
-        if (Platform.isWindows) {
-          final shortcuts = {
-            const SingleActivator(LogicalKeyboardKey.keyV):
-                VoidCallbackIntent(context.read<ToolCubit>().setMove),
-            const SingleActivator(LogicalKeyboardKey.keyF):
-                VoidCallbackIntent(context.read<ToolCubit>().setFrame),
-            const SingleActivator(LogicalKeyboardKey.keyR):
-                VoidCallbackIntent(context.read<ToolCubit>().setRectangle),
-            const SingleActivator(LogicalKeyboardKey.keyH):
-                VoidCallbackIntent(context.read<ToolCubit>().setHand),
-            const SingleActivator(LogicalKeyboardKey.keyT):
-                VoidCallbackIntent(context.read<ToolCubit>().setText),
-            const SingleActivator(LogicalKeyboardKey.delete):
-                VoidCallbackIntent(selected.isEmpty
-                    ? () {}
-                    : () => componentsNotifier.removeSelected()),
-            const SingleActivator(LogicalKeyboardKey.bracketLeft):
-                VoidCallbackIntent(
-                    selected.isEmpty ? () {} : () => handleGoBackward()),
-            const SingleActivator(LogicalKeyboardKey.bracketRight):
-                VoidCallbackIntent(
-                    selected.isEmpty ? () {} : () => handleGoForward()),
-          };
-          var i = 0;
-          while (i < shortcuts.length) {
-            if (ShortcutActivator.isActivatedBy(
-                shortcuts.keys.elementAt(i), event)) {
-              shortcuts.values.elementAt(i).callback();
-            }
-            i++;
-          }
-        }
-      },
-      child: Scaffold(
-        backgroundColor: const Color(0xFF333333),
-        body: Stack(
-          clipBehavior: Clip.none,
-          children: [
-            /// Canvas Background
-            ValueListenableBuilder(
-              valueListenable: transformationController,
-              builder: (context, transform, child) {
-                return Positioned(
-                  left: transform.getTranslation().x + sidebarWidth,
-                  top: transform.getTranslation().y + topbarHeight,
-                  child: Container(
-                    width: canvasStateProvider.size.width *
-                        transform.getMaxScaleOnAxis(),
-                    height: canvasStateProvider.size.height *
-                        transform.getMaxScaleOnAxis(),
-                    foregroundDecoration: BoxDecoration(
-                      border: !canvasStateProvider.hidden
-                          ? null
-                          : Border.all(
-                              color: Colors.grey,
-                              strokeAlign: BorderSide.strokeAlignOutside,
-                              width: 2,
-                            ),
-                      borderRadius: BorderRadius.circular(1),
-                    ),
-                    decoration: BoxDecoration(
-                      color: canvasStateProvider.hidden
-                          ? null
-                          : canvasStateProvider.color,
-                      border: !canvasStateProvider.hidden
-                          ? null
-                          : Border.all(
-                              color: Colors.grey,
-                              strokeAlign: BorderSide.strokeAlignOutside,
-                              width: 1),
-                      boxShadow: const [
-                        BoxShadow(
-                          color: Colors.black45,
-                          blurRadius: 24,
-                          blurStyle: BlurStyle.outer,
-                        ),
-                      ],
-                    ),
-                  ),
-                );
-              },
-            ),
-
-            /// Components and Controllers
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: sidebarWidth)
-                  .copyWith(top: topbarHeight),
-              // Canvas
-              child: Stack(
-                children: [
-                  // Background/canvas
-                  ValueListenableBuilder(
-                    valueListenable: transformationController,
-                    builder: (context, transform, child) => Stack(
-                      clipBehavior: Clip.none,
-                      children: [
-                        // unselect all
-                        GestureDetector(
-                          onTap: () {
-                            selectedNotifier.clear();
-                            hoveredNotifier.clear();
-                            isShowColorPicker = false;
-                            setState(() {});
-                          },
-                          child: Container(
-                            width: mqSize.width,
-                            height: mqSize.height,
-                            color: Colors.transparent,
-                          ),
-                        ),
-                        // components label
-                        ...components
-                            .where((element) =>
-                                element.type == ComponentType.frame)
-                            .map(
-                          (e) {
-                            return buildComponentLabel(
-                              e,
-                              canvasStateProvider.color,
-                            );
-                          },
-                        ),
-
-                        // components
-                        ValueListenableBuilder(
-                          valueListenable: transformationController,
-                          builder: (context, transform, child) => Stack(
-                            clipBehavior: Clip.none,
-                            children: [
-                              ...components.mapIndexed(
-                                (i, component) {
-                                  return component.hidden ||
-                                          (component.type ==
-                                                  ComponentType.text &&
-                                              selected.contains(i) &&
-                                              canvasEvents.containsAny(
-                                                  {CanvasEvent.editingText}))
-                                      ? const SizedBox.shrink()
-                                      : buildComponentWidget(
-                                          i, component, transform);
-                                },
-                              ),
-                            ],
-                          ),
-                        ),
-
-                        // controller for selections
-                        AnimatedBuilder(
-                          animation: Listenable.merge([
-                            componentsNotifier.state,
-                            selectedNotifier.state,
-                          ]),
-                          builder: (context, child) => Stack(
-                            children: componentsNotifier.state.value
-                                .mapIndexed((i, e) => ControllerWidget(i))
-                                .toList(),
-                          ),
-                        ),
-
-                        // selected controller(s) (should be at the front of every
-                        // controllers)
-                        AnimatedBuilder(
-                          animation: Listenable.merge([
-                            componentsNotifier.state,
-                            selectedNotifier.state,
-                          ]),
-                          builder: (context, child) => Stack(
-                            children: selectedNotifier.state.value
-                                .mapIndexed((i, e) => ControllerWidget(e))
-                                .toList(),
-                          ),
-                        ),
-                        // components
-                        if (canvasEvents.containsAny({CanvasEvent.editingText}))
-                          ValueListenableBuilder(
-                            valueListenable: transformationController,
-                            builder: (context, transform, child) => Stack(
-                              clipBehavior: Clip.none,
-                              children: [
-                                ...selectedNotifier.state.value
-                                    .where((e) =>
-                                        components[e].type ==
-                                        ComponentType.text)
-                                    .mapIndexed(
-                                  (i, e) {
-                                    final component = components[e];
-
-                                    return component.hidden
-                                        ? const SizedBox.shrink()
-                                        : buildComponentWidget(
-                                            e, component, transform,
-                                            editText: true);
-                                  },
-                                ),
-                              ],
-                            ),
-                          ),
-
-                        if (kDebugMode)
-                          ...context
-                              .watch<DebugPointCubit>()
-                              .state
-                              .map((e) => Builder(builder: (context) {
-                                    final component = e;
-                                    final scale = transform.getMaxScaleOnAxis();
-                                    return Positioned(
-                                      left: transform.getTranslation().x +
-                                          component.dx * scale,
-                                      top: transform.getTranslation().y +
-                                          component.dy * scale,
-                                      child: Container(
-                                        width: 5,
-                                        height: 5,
-                                        color: Colors.red,
-                                      ),
-                                    );
-                                  }))
-                      ],
-                    ),
-                  ),
-
-                  //
-
-                  if (tool == ToolType.rectangle ||
-                      isCreateTooling ||
-                      tool == ToolType.frame ||
-                      tool == ToolType.text)
-                    const CreatorWidget(),
-
-                  // camera (kinda)
-                  TransparentPointer(
-                    transparent: !isToolHand,
-                    child: Listener(
-                      onPointerDown: (event) {
-                        final state = (event.kind == PointerDeviceKind.mouse &&
-                                event.buttons == kMiddleMouseButton)
-                            ? CanvasEvent.middleClick
-                            : CanvasEvent.leftClick;
-                        context.read<CanvasEventsCubit>().add(
-                              state,
-                            );
-                      },
-                      onPointerUp: (event) {
-                        (event.kind == PointerDeviceKind.mouse &&
-                                event.buttons == kMiddleMouseButton)
-                            ? CanvasEvent.middleClick
-                            : CanvasEvent.leftClick;
-                        context.read<CanvasEventsCubit>().remove(
-                              CanvasEvent.leftClick,
-                            );
-                      },
-                      child: MouseRegion(
-                        cursor: switch ((
-                          context
-                              .read<CanvasEventsCubit>()
-                              .state
-                              .contains(CanvasEvent.leftClick),
-                          isToolHand && (!isNotCanvasTooling || isCanvasTooling)
-                        )) {
-                          (false, true) => SystemMouseCursors.grab,
-                          (true, true) => SystemMouseCursors.grabbing,
-                          _ => MouseCursor.defer
-                        },
-                        child: InteractiveViewer.builder(
-                          transformationController: transformationController,
-                          panEnabled: (isToolHand &&
-                                  (!isNotCanvasTooling || isCanvasTooling)) ||
-                              (isCanvasTooling && !leftClick && !isZooming),
-                          onInteractionStart: (details) {
-                            if (pressedMeta) {
-                              context.read<CanvasEventsCubit>().add(
-                                    CanvasEvent.zoomingCanvas,
-                                  );
-                            } else if (isToolHand || middleClick) {
-                              context.read<CanvasEventsCubit>().add(
-                                    CanvasEvent.panningCanvas,
-                                  );
-                            }
-                          },
-                          onInteractionUpdate: (details) {
-                            context
-                                .read<PointerCubit>()
-                                .update(details.focalPoint);
-                            context
-                                .read<CanvasTransformCubit>()
-                                .update(transformationController.value);
-                            if (details.scale == 1 && !pressedMeta) {
-                              context.read<CanvasEventsCubit>().add(
-                                    CanvasEvent.panningCanvas,
-                                  );
-                            }
-                          },
-                          onInteractionEnd: (details) =>
-                              context.read<CanvasEventsCubit>()
-                                ..remove(CanvasEvent.panningCanvas)
-                                ..remove(CanvasEvent.zoomingCanvas),
-                          maxScale: 256,
-                          minScale: .01,
-                          trackpadScrollCausesScale:
-                              (pressedMeta || isZooming) &&
-                                  !canvasEvents
-                                      .containsAny({CanvasEvent.panningCanvas}),
-                          interactionEndFrictionCoefficient: 0.000135,
-                          boundaryMargin: const EdgeInsets.all(double.infinity),
-                          builder: (context, viewport) => const SizedBox(),
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-
-            /// Top bar, side bars
-            Column(
-              children: [
-                /// Top bar
-                TopBar(),
-                Expanded(
-                  child: Row(
-                    children: [
-                      /// Left Sidebar
-                      LeftSidebar(),
-                      const Expanded(
-                        child: SizedBox.shrink(),
-                      ),
-                      RightSidebar(
-                        toggleColorPicker: () {
-                          isShowColorPicker = !isShowColorPicker;
-                          setState(() {});
-                        },
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-
-            /// Color picker
-            if (isShowColorPicker)
-              Positioned(
-                top: topbarHeight + 4,
-                right: sidebarWidth + 4,
-                child: Container(
-                  width: 240,
-                  height: 320,
-                  decoration: BoxDecoration(
-                    color: colorScheme.surface,
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Column(
-                      children: [
-                        Expanded(
-                          child: ColorPicker(
-                            color: canvasStateProvider.color,
-                            onColorChanged: (value) {
-                              canvasStateNotifier.update(
-                                backgroundColor: value,
-                              );
-                            },
-                            padding: EdgeInsets.zero,
-                            wheelSquarePadding: 24,
-                            columnSpacing: 0,
-                            opacityTrackHeight: 10,
-                            opacityThumbRadius: 12,
-                            wheelWidth: 8,
-                            colorCodeHasColor: true,
-                            showColorCode: true,
-                            enableOpacity: true,
-                            enableShadesSelection: false,
-                            enableTonalPalette: false,
-                            pickersEnabled: const {
-                              ColorPickerType.wheel: true,
-                              ColorPickerType.accent: false,
-                              ColorPickerType.primary: false,
-                              ColorPickerType.both: false,
-                            },
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-
-            /// Custom pointer handler
-            Positioned.fill(
-              child: TransparentPointer(
-                child: Builder(
-                  builder: (context) {
-                    return MouseRegion(
-                      cursor: switch (leftClick) {
-                        true
-                            when context.read<CanvasEventsCubit>().containsAny(
-                              {
-                                CanvasEvent.resizingComponent,
-                                CanvasEvent.rotatingComponent
-                              },
-                            ) =>
-                          SystemMouseCursors.none,
-                        _ => MouseCursor.defer
-                      },
-                      onHover: (event) =>
-                          context.read<PointerCubit>().update(event.position),
-                    );
-                  },
-                ),
-              ),
-            ),
-
-            /// Custom pointer
-            AnimatedBuilder(
-              animation: Listenable.merge([
-                componentsNotifier.state,
-                selectedNotifier.state,
-              ]),
-              builder: (context, child) {
-                final stateContains =
-                    context.read<CanvasEventsCubit>().containsAny;
-
-                final isRotating = context
-                    .read<CanvasEventsCubit>()
-                    .state
-                    .contains(CanvasEvent.rotatingComponent);
-
-                final enabled = stateContains({
-                  CanvasEvent.rotatingComponent,
-                  CanvasEvent.resizingComponent
-                });
-
-                var angle = 0.0;
-                if (selectedNotifier.state.value.isNotEmpty && enabled) {
-                  angle = componentsNotifier
-                          .state
-                          .value[selectedNotifier.state.value.first]
-                          .component
-                          .angle +
-                      (isRotating ? pi / 6 : 0);
-                }
-
-                return !enabled
-                    ? const SizedBox.shrink()
-                    : Transform.translate(
-                        offset: context.read<PointerCubit>().state,
-                        child: Transform.translate(
-                          offset: const Offset(-12, -12),
-                          child: Transform.rotate(
-                            angle: angle,
-                            child: IgnorePointer(
-                              child: SizedBox(
-                                width: 24,
-                                height: 24,
-                                child: Icon(
-                                  isRotating
-                                      ? CupertinoIcons.arrow_turn_up_right
-                                      : CupertinoIcons.arrow_left_right,
-                                  size: 14,
-                                  color: colorScheme.onPrimary,
-                                ),
-                              ),
-                            ),
-                          ),
-                        ),
-                      );
-              },
-            )
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget buildComponentLabel(
-    ComponentData e,
-    Color backgroundColor,
-  ) {
-    return ValueListenableBuilder(
-      valueListenable: context.read<CanvasTransformCubit>().state,
-      builder: (context, matrix, child) {
-        final component = e.component;
-        final tWidth = component.size.width;
-        final tHeight = component.size.height;
-
-        final edges = component.rotatedEdges;
-        final topLeft = edges.tl;
-        final topRight = edges.tr;
-        final bottomLeft = edges.bl;
-        final bottomRight = edges.br;
-
-        var edge = switch ((tWidth < 0, tHeight < 0)) {
-          (true, false) => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => bottomLeft,
-              < 135 => topLeft,
-              < 225 => topRight,
-              < 315 => bottomRight,
-              < 360 => bottomLeft,
-              _ => topLeft,
-            },
-          (true, true) => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => topLeft,
-              < 135 => bottomLeft,
-              < 225 => bottomRight,
-              < 315 => topRight,
-              < 360 => topLeft,
-              _ => topLeft,
-            },
-          (false, true) => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => bottomRight,
-              < 135 => bottomLeft,
-              < 225 => topLeft,
-              < 315 => topRight,
-              < 360 => bottomRight,
-              _ => topLeft,
-            },
-          _ => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => topRight,
-              < 135 => topLeft,
-              < 225 => bottomLeft,
-              < 315 => bottomRight,
-              < 360 => topRight,
-              _ => topLeft,
-            }
-        };
-        final newAngle = switch ((tWidth < 0, tHeight < 0)) {
-          (true, false) => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => pi / 2,
-              < 135 => 0,
-              < 225 => -pi / 2,
-              < 315 => pi,
-              < 360 => pi / 2,
-              _ => 0.0,
-            },
-          (true, true) => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => pi / 2,
-              < 135 => 0,
-              < 225 => -pi / 2,
-              < 315 => pi,
-              < 360 => pi / 2,
-              _ => 0.0,
-            },
-          _ => switch ((component.angle / pi * 180 + 90) % 360) {
-              < 45 => pi / 2,
-              < 135 => 0,
-              < 225 => -pi / 2,
-              < 315 => pi,
-              < 360 => pi / 2,
-              _ => 0.0,
-            }
-        };
-
-        edge = MatrixUtils.transformPoint(matrix, edge);
-
-        final scale = matrix.getMaxScaleOnAxis();
-
-        return Positioned(
-          left: edge.dx,
-          top: edge.dy,
-          child: Transform.rotate(
-            angle: component.angle + newAngle,
-            alignment: Alignment.topLeft,
-            child: AnimatedSlide(
-              offset: Offset(tWidth < 0 ? -1 : 0, -1),
-              duration: Duration.zero,
-              child: SizedBox(
-                width: (tWidth * scale).abs(),
-                height: 24,
-                child: Text(
-                  e.name,
-                  overflow: TextOverflow.ellipsis,
-                  style: TextStyle(
-                    color: backgroundColor.computeLuminance() > 0.5
-                        ? Colors.black.withOpacity(.5)
-                        : Colors.white.withOpacity(.5),
-                  ),
-                ),
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget buildComponentWidget(int index, ComponentData e, Matrix4 transform,
-      {bool editText = false}) {
-    final component = e.component;
-    final scale = transform.getMaxScaleOnAxis();
-    final tWidth = component.size.width;
-    final tHeight = component.size.height;
-
-    final child = e.type == ComponentType.text
-        ? SizedBox(
-            width: tWidth < 0 ? -tWidth : tWidth,
-            height: tHeight < 0 ? -tHeight : tHeight,
-            child: !editText
-                ? Text(
-                    e.name,
-                    style: Theme.of(context).textTheme.bodyMedium,
-                  )
-                : TextField(
-                    autofocus: true,
-                    controller: e.textController,
-                    onChanged: (value) {
-                      componentsNotifier.replace(index, name: value);
-
-                      final textStyle = Theme.of(context).textTheme.bodyMedium;
-
-                      final textSpan = TextSpan(
-                        text: value,
-                        style: textStyle,
-                      );
-
-                      final textPainter = TextPainter(
-                        text: textSpan,
-                        textDirection: TextDirection.ltr,
-                      );
-
-                      textPainter.layout(
-                        minWidth: 0,
-                        maxWidth: tWidth.abs(),
-                      );
-
-                      final newHeight = textPainter.height;
-
-                      if (newHeight != component.size.height) {
-                        final newRect = rotateRect(
-                          Rect.fromLTWH(
-                            component.rect.topLeft.dx,
-                            component.rect.topLeft.dy,
-                            component.size.width,
-                            newHeight,
-                          ),
-                          component.angle,
-                          component.rect.center,
-                        );
-
-                        componentsNotifier.replace(
-                          index,
-                          transform: Component(
-                            newRect.unrotated.topLeft,
-                            Size(component.size.width, newHeight),
-                            component.angle,
-                            component.flipX,
-                            component.flipY,
-                            // keepOrigin: true,
-                          ),
-                        );
-                      }
-                    },
-                    onTapOutside: (event) {
-                      print('TAPOUTSIDE');
-                      context
-                          .read<CanvasEventsCubit>()
-                          .remove(CanvasEvent.editingText);
-                      e.textController?.dispose();
-                    },
-                    expands: true,
-                    maxLines: null,
-                    style: Theme.of(context).textTheme.bodyMedium,
-                    cursorHeight:
-                        Theme.of(context).textTheme.bodyMedium?.fontSize,
-                    decoration: const InputDecoration(
-                      border: InputBorder.none,
-                      contentPadding: EdgeInsets.symmetric(vertical: 3),
-                    ),
-                  ),
-          )
-        : Container(
-            width: tWidth < 0 ? -tWidth : tWidth,
-            height: tHeight < 0 ? -tHeight : tHeight,
-            decoration: BoxDecoration(
-              borderRadius: e.borderRadius,
-              border: e.border,
-              color: e.color,
-            ),
-          );
-
-    return Positioned(
-      left: transform.getTranslation().x +
-          (component.pos.dx + (tWidth < 0 ? tWidth : 0)) * scale,
-      top: transform.getTranslation().y +
-          (component.pos.dy + (tHeight < 0 ? tHeight : 0)) * scale,
-      child: Transform.scale(
-        scale: scale,
-        alignment: Alignment.topLeft,
-        child: Transform.rotate(
-          angle: component.angle,
-          child: Transform.flip(
-            flipX: component.flipX,
-            flipY: component.flipY,
-            child: child,
-          ),
-        ),
-      ),
-    );
-  }
 }
